+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [v6.11.1] 2022-06-28
 ### Fixed
 - [#88](https://github.com/seznam/slo-exporter/pull/88) slo-rules-generator does not output empty `escalate` label
 - [#87](https://github.com/seznam/slo-exporter/pull/87) slo-exporter's SLO recording rule for uptime success events
+- slo-rules-generator now properly truncates files when overwriting them
 
 ## [v6.11.0] 2022-01-19
 ### Added